Deleted portfolio page and migrated portfolio contents into article.
So I decided to delete the portfolio page and only left the article page
on the blog. Well the reason is that I think that it's somewhat
redundant to what I had currently. I already have an portfolio site
(still under development) which will be showcasing the projects that
I've worked on.